+++ airflow/auth/managers/fab/api/auth/backend/kerberos_auth.py
@@ -15,25 +15,28 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""
+This module is deprecated.
+
+Please use :mod:`airflow.providers.fab.auth_manager.api.auth.backend.kerberos_auth` instead.
+"""
 from __future__ import annotations
 
-import logging
-from functools import partial
+import warnings
 from typing import Any
 
 from requests_kerberos import HTTPKerberosAuth
 
-from airflow.api.auth.backend.kerberos_auth import (
-    init_app as base_init_app,
-    requires_authentication as base_requires_authentication,
-)
-from airflow.utils.airflow_flask_app import get_airflow_app
-
-log = logging.getLogger(__name__)
+import airflow.providers.fab.auth_manager.api.auth.backend.kerberos_auth as fab_kerberos_auth
+from airflow.exceptions import RemovedInAirflow3Warning
 
 CLIENT_AUTH: tuple[str, str] | Any | None = HTTPKerberosAuth(service="airflow")
 
-init_app = base_init_app
-requires_authentication = partial(
-    base_requires_authentication, find_user=get_airflow_app().appbuilder.sm.find_user
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.fab.auth_manager.api.auth.backend.kerberos_auth` instead.",
+    RemovedInAirflow3Warning,
+    stacklevel=2,
 )
+
+init_app = fab_kerberos_auth.init_app
+requires_authentication = fab_kerberos_auth.requires_authentication