+++ airflow/providers/cncf/kubernetes/hooks/kubernetes.py
@@ -86,8 +86,8 @@ class KubernetesHook(BaseHook, PodOperatorHookProtocol):
 
     DEFAULT_NAMESPACE = "default"
 
-    @staticmethod
-    def get_connection_form_widgets() -> dict[str, Any]:
+    @classmethod
+    def get_connection_form_widgets(cls) -> dict[str, Any]:
         """Return connection widgets to add to connection form."""
         from flask_appbuilder.fieldwidgets import BS3TextFieldWidget
         from flask_babel import lazy_gettext
@@ -111,8 +111,8 @@ class KubernetesHook(BaseHook, PodOperatorHookProtocol):
             ),
         }
 
-    @staticmethod
-    def get_ui_field_behaviour() -> dict[str, Any]:
+    @classmethod
+    def get_ui_field_behaviour(cls) -> dict[str, Any]:
         """Return custom field behaviour."""
         return {
             "hidden_fields": ["host", "schema", "login", "password", "port", "extra"],