+++ airflow/serialization/pydantic/dag_run.py
@@ -55,6 +55,7 @@ class DagRunPydantic(BaseModelPydantic):
     updated_at: Optional[datetime]
     dag: Optional[PydanticDag]
     consumed_dataset_events: List[DatasetEventPydantic]  # noqa
+    log_template_id: Optional[int]
 
     model_config = ConfigDict(from_attributes=True, arbitrary_types_allowed=True)