+++ airflow/serialization/pydantic/taskinstance.py
@@ -87,7 +87,6 @@ class TaskInstancePydantic(BaseModelPydantic, LoggingMixin):
     pool_slots: int
     queue: str
     priority_weight: Optional[int]
-    priority_weight_strategy: Optional[str]
     operator: str
     custom_operator_name: Optional[str]
     queued_dttm: Optional[str]