+++ airflow/utils/db.py
@@ -88,7 +88,7 @@ _REVISION_HEADS_MAP = {
     "2.6.0": "98ae134e6fff",
     "2.6.2": "c804e5c76e3e",
     "2.7.0": "405de8318b3a",
-    "2.8.0": "624ecf3b6a5e",
+    "2.8.0": "10b52ebd31f7",
 }
 
 
@@ -703,8 +703,8 @@ def _get_flask_db(sql_database_uri):
 def _create_db_from_orm(session):
     from alembic import command
 
-    from airflow.auth.managers.fab.models import Model
     from airflow.models.base import Base
+    from airflow.providers.fab.auth_manager.models import Model
 
     def _create_flask_session_tbl(sql_database_uri):
         db = _get_flask_db(sql_database_uri)
@@ -983,7 +983,7 @@ def check_username_duplicates(session: Session) -> Iterable[str]:
     :param session:  session of the sqlalchemy
     :rtype: str
     """
-    from airflow.auth.managers.fab.models import RegisterUser, User
+    from airflow.providers.fab.auth_manager.models import RegisterUser, User
 
     for model in [User, RegisterUser]:
         dups = []
@@ -1712,8 +1712,8 @@ def drop_airflow_models(connection):
     :param connection: SQLAlchemy Connection
     :return: None
     """
-    from airflow.auth.managers.fab.models import Model
     from airflow.models.base import Base
+    from airflow.providers.fab.auth_manager.models import Model
 
     Base.metadata.drop_all(connection)
     Model.metadata.drop_all(connection)