+++ dev/breeze/src/airflow_breeze/commands/production_image_commands_config.py
@@ -36,6 +36,7 @@ PRODUCTION_IMAGE_TOOLS_PARAMETERS: dict[str, list[dict[str, str | list[str]]]] =
                 "--docker-cache",
                 "--version-suffix-for-pypi",
                 "--build-progress",
+                "--docker-host",
             ],
         },
         {
@@ -91,6 +92,7 @@ PRODUCTION_IMAGE_TOOLS_PARAMETERS: dict[str, list[dict[str, str | list[str]]]] =
                 "--installation-method",
                 "--install-airflow-reference",
                 "--install-packages-from-context",
+                "--install-mysql-client-type",
                 "--cleanup-context",
                 "--use-constraints-for-context-packages",
                 "--disable-airflow-repo-cache",