+++ dev/breeze/src/airflow_breeze/params/shell_params.py
@@ -20,7 +20,6 @@ import os
 from copy import deepcopy
 from dataclasses import dataclass, field
 from functools import cached_property
-from os import _Environ
 from pathlib import Path
 
 from airflow_breeze.branch_defaults import AIRFLOW_BRANCH, DEFAULT_AIRFLOW_CONSTRAINTS_BRANCH
@@ -42,7 +41,6 @@ from airflow_breeze.global_constants import (
     MOUNT_ALL,
     MOUNT_REMOVE,
     MOUNT_SELECTED,
-    MOUNT_SKIP,
     MSSQL_HOST_PORT,
     MYSQL_HOST_PORT,
     POSTGRES_HOST_PORT,
@@ -54,8 +52,8 @@ from airflow_breeze.global_constants import (
     get_airflow_version,
 )
 from airflow_breeze.utils.console import get_console
-from airflow_breeze.utils.host_info_utils import get_host_group_id, get_host_os
-from airflow_breeze.utils.packages import get_suspended_provider_folders
+from airflow_breeze.utils.docker_command_utils import is_docker_rootless
+from airflow_breeze.utils.host_info_utils import get_host_group_id, get_host_os, get_host_user_id
 from airflow_breeze.utils.path_utils import (
     AIRFLOW_SOURCES_ROOT,
     BUILD_CACHE_DIR,
@@ -78,6 +76,26 @@ def add_mssql_compose_file(compose_file_list: list[Path]):
         compose_file_list.append(DOCKER_COMPOSE_DIR / "backend-mssql-tmpfs-volume.yml")
     else:
         compose_file_list.append(DOCKER_COMPOSE_DIR / "backend-mssql-docker-volume.yml")
+
+
+def generated_socket_compose_file(local_socket_path: str) -> str:
+    return f"""
+---
+services:
+  airflow:
+    volumes:
+      - {local_socket_path}:/var/run/docker.sock
+"""
+
+
+def generated_docker_host_environment() -> str:
+    return """
+---
+services:
+  airflow:
+    environment:
+      - DOCKER_HOST=${DOCKER_HOST}
+"""
 
 
 def _set_var(env: dict[str, str], variable: str, attribute: str | bool | None, default: str | None = None):
@@ -110,14 +128,17 @@ class ShellParams:
     """
 
     airflow_branch: str = os.environ.get("DEFAULT_BRANCH", AIRFLOW_BRANCH)
+    airflow_constraints_location: str = ""
     airflow_constraints_mode: str = ALLOWED_CONSTRAINTS_MODES_CI[0]
-    airflow_constraints_reference: str = DEFAULT_AIRFLOW_CONSTRAINTS_BRANCH
+    airflow_constraints_reference: str = ""
     airflow_extras: str = ""
+    airflow_skip_constraints: bool = False
     backend: str = ALLOWED_BACKENDS[0]
     base_branch: str = "main"
     builder: str = "autodetect"
     celery_broker: str = DEFAULT_CELERY_BROKER
     celery_flower: bool = False
+    chicken_egg_providers: str = ""
     collect_only: bool = False
     database_isolation: bool = False
     db_reset: bool = False
@@ -125,13 +146,14 @@ class ShellParams:
         "DEFAULT_CONSTRAINTS_BRANCH", DEFAULT_AIRFLOW_CONSTRAINTS_BRANCH
     )
     dev_mode: bool = False
+    docker_host: str | None = os.environ.get("DOCKER_HOST")
     downgrade_sqlalchemy: bool = False
     dry_run: bool = False
     enable_coverage: bool = False
     executor: str = START_AIRFLOW_DEFAULT_ALLOWED_EXECUTOR
     extra_args: tuple = ()
     force_build: bool = False
-    forward_credentials: str = "false"
+    forward_credentials: bool = False
     forward_ports: bool = True
     github_actions: str = os.environ.get("GITHUB_ACTIONS", "false")
     github_repository: str = APACHE_AIRFLOW_GITHUB_REPOSITORY
@@ -157,6 +179,10 @@ class ShellParams:
     platform: str = DOCKER_DEFAULT_PLATFORM
     postgres_version: str = ALLOWED_POSTGRES_VERSIONS[0]
     project_name: str = ALLOWED_DOCKER_COMPOSE_PROJECTS[0]
+    providers_constraints_location: str = ""
+    providers_constraints_mode: str = ALLOWED_CONSTRAINTS_MODES_CI[0]
+    providers_constraints_reference: str = ""
+    providers_skip_constraints: bool = False
     python: str = ALLOWED_PYTHON_MAJOR_MINOR_VERSIONS[0]
     quiet: bool = False
     regenerate_missing_docs: bool = False
@@ -165,7 +191,6 @@ class ShellParams:
     run_db_tests_only: bool = False
     run_system_tests: bool = os.environ.get("RUN_SYSTEM_TESTS", "false") == "true"
     run_tests: bool = False
-    skip_constraints: bool = False
     skip_db_tests: bool = False
     skip_environment_initialization: bool = False
     skip_image_upgrade_check: bool = False
@@ -173,7 +198,7 @@ class ShellParams:
     skip_provider_tests: bool = False
     skip_ssh_setup: bool = os.environ.get("SKIP_SSH_SETUP", "false") == "true"
     standalone_dag_processor: bool = False
-    start_airflow: str = "false"
+    start_airflow: bool = False
     test_type: str | None = None
     tty: str = "auto"
     upgrade_boto: bool = False
@@ -273,7 +298,13 @@ class ShellParams:
             get_console().print(f"[info]Airflow used at runtime: {self.use_airflow_version}[/]")
 
     def get_backend_compose_files(self, backend: str) -> list[Path]:
-        backend_docker_compose_file = DOCKER_COMPOSE_DIR / f"backend-{backend}.yml"
+        if backend == "sqlite" and self.project_name != "breeze":
+            # When running scripts, we do not want to mount the volume to make sure that the
+            # sqlite database is not persisted between runs of the script and that the
+            # breeze database is not cleaned accidentally
+            backend_docker_compose_file = DOCKER_COMPOSE_DIR / f"backend-{backend}-no-volume.yml"
+        else:
+            backend_docker_compose_file = DOCKER_COMPOSE_DIR / f"backend-{backend}.yml"
         if backend in ("sqlite", "none") or not self.forward_ports:
             return [backend_docker_compose_file]
         if self.project_name == "pre-commit":
@@ -299,19 +330,14 @@ class ShellParams:
             compose_file_list.append(DOCKER_COMPOSE_DIR / "integration-celery.yml")
 
         compose_file_list.append(DOCKER_COMPOSE_DIR / "base.yml")
+        self.add_docker_in_docker(compose_file_list)
         compose_file_list.extend(backend_files)
         compose_file_list.append(DOCKER_COMPOSE_DIR / "files.yml")
 
-        if self.image_tag is not None and self.image_tag != "latest":
-            get_console().print(
-                f"[warning]Running tagged image tag = {self.image_tag}. "
-                f"Forcing mounted sources to be 'skip'[/]"
-            )
-            self.mount_sources = MOUNT_SKIP
         if self.use_airflow_version is not None:
             get_console().print(
-                "[info]Forcing --mount-sources to `remove` since we are not installing airflow "
-                f"from sources but from {self.use_airflow_version}[/]"
+                "\n[warning]Forcing --mount-sources to `remove` since we are not installing airflow "
+                f"from sources but from {self.use_airflow_version}[/]\n"
             )
             self.mount_sources = MOUNT_REMOVE
         if self.forward_ports and not self.project_name == "pre-commit":
@@ -324,8 +350,6 @@ class ShellParams:
             compose_file_list.append(DOCKER_COMPOSE_DIR / "remove-sources.yml")
         if self.forward_credentials:
             compose_file_list.append(DOCKER_COMPOSE_DIR / "forward-credentials.yml")
-        if self.use_airflow_version is not None:
-            compose_file_list.append(DOCKER_COMPOSE_DIR / "remove-sources.yml")
         if self.include_mypy_volume:
             compose_file_list.append(DOCKER_COMPOSE_DIR / "mypy.yml")
         if "all-testable" in self.integration:
@@ -363,6 +387,8 @@ class ShellParams:
 
     @cached_property
     def suspended_providers_folders(self):
+        from airflow_breeze.utils.packages import get_suspended_provider_folders
+
         return " ".join(get_suspended_provider_folders()).strip()
 
     @cached_property
@@ -381,25 +407,62 @@ class ShellParams:
             # mssql_data_volume variable is only used in case of tmpfs
             return ""
 
-    @cached_property
-    def rootless_docker(self) -> bool:
-        try:
-            response = run_command(
-                ["docker", "info", "-f", "{{println .SecurityOptions}}"],
-                capture_output=True,
-                check=False,
-                text=True,
-            )
-            if response.returncode == 0 and "rootless" in response.stdout.strip():
-                get_console().print("[info]Docker is running in rootless mode.[/]\n")
-                return True
-        except FileNotFoundError:
-            # we ignore if docker is missing
-            pass
-        return False
+    def add_docker_in_docker(self, compose_file_list: list[Path]):
+        generated_compose_file = DOCKER_COMPOSE_DIR / "_generated_docker_in_docker.yml"
+        unix_prefix = "unix://"
+        if self.docker_host:
+            if self.docker_host.startswith(unix_prefix):
+                # Socket is locally available
+                socket_path = Path(self.docker_host[len(unix_prefix) :])
+                if (
+                    get_host_os() == "darwin"
+                    and socket_path.resolve() == (Path.home() / ".docker" / "run" / "docker.sock").resolve()
+                ):
+                    # We are running on MacOS and the socket is the default "user" bound one
+                    # We need to pretend that we are running on Linux and use the default socket
+                    # in the VM instead - see https://github.com/docker/for-mac/issues/6545
+                    compose_file_list.append(DOCKER_COMPOSE_DIR / "docker-socket.yml")
+                    return
+                if socket_path.is_socket():
+                    generated_compose_file.write_text(generated_socket_compose_file(socket_path.as_posix()))
+                    compose_file_list.append(generated_compose_file)
+                else:
+                    get_console().print(
+                        f"[warning]The socket {socket_path} pointed at by DOCKER_HOST does not exist or is "
+                        "not a socket. Cannot use it for docker-compose for docker-in-docker forwarding[/]\n"
+                        "[info]If you know where your socket is, you can set DOCKER_HOST "
+                        "environment variable to unix://path_to_docker.sock[/]\n"
+                    )
+            else:
+                # Socket is something different (TCP?) just pass it through as DOCKER_HOST variable
+                generated_compose_file.write_text(generated_docker_host_environment())
+                compose_file_list.append(generated_compose_file)
+        elif self.rootless_docker:
+            xdg_runtime_dir = Path(os.environ.get("XDG_RUNTIME_DIR", f"/run/user/{get_host_user_id()}"))
+            socket_path = xdg_runtime_dir / "docker.sock"
+            if socket_path.is_socket():
+                generated_compose_file.write_text(generated_socket_compose_file(socket_path.as_posix()))
+                compose_file_list.append(generated_compose_file)
+            else:
+                get_console().print(
+                    f"[warning]The socket {socket_path} does not exist or is not a socket. "
+                    "Cannot use it for docker-compose for docker-in-docker forwarding[/]\n"
+                    "[info]If you know where your socket is, you can set DOCKER_HOST environment variable "
+                    "to unix://path_to_docker.sock[/]\n"
+                )
+        else:
+            # We fall back to default docker socket when no host is defined including MacOS
+            # NOTE! Even if we are using "desktop-linux" context where "/var/run/docker.sock" is not used,
+            # Docker engine works fine because "/var/run/docker.sock" is mounted at the VM and there
+            # the /var/run/docker.sock is available. See https://github.com/docker/for-mac/issues/6545
+            compose_file_list.append(DOCKER_COMPOSE_DIR / "docker-socket.yml")
 
     @cached_property
-    def env_variables_for_docker_commands(self) -> _Environ:
+    def rootless_docker(self) -> bool:
+        return is_docker_rootless()
+
+    @cached_property
+    def env_variables_for_docker_commands(self) -> dict[str, str]:
         """
         Constructs environment variables needed by the docker-compose command, based on Shell parameters
         passed to it.
@@ -413,18 +476,13 @@ class ShellParams:
         _env: dict[str, str] = {}
         _set_var(_env, "AIRFLOW_CI_IMAGE", self.airflow_image_name)
         _set_var(_env, "AIRFLOW_CI_IMAGE_WITH_TAG", self.airflow_image_name_with_tag)
-        _set_var(
-            _env, "AIRFLOW_CONSTRAINTS_MODE", self.airflow_constraints_mode, "constraints-source-providers"
-        )
-        _set_var(
-            _env,
-            "AIRFLOW_CONSTRAINTS_REFERENCE",
-            self.airflow_constraints_reference,
-            "constraints-source-providers",
-        )
+        _set_var(_env, "AIRFLOW_CONSTRAINTS_LOCATION", self.airflow_constraints_location)
+        _set_var(_env, "AIRFLOW_CONSTRAINTS_MODE", self.airflow_constraints_mode)
+        _set_var(_env, "AIRFLOW_CONSTRAINTS_REFERENCE", self.airflow_constraints_reference)
         _set_var(_env, "AIRFLOW_ENABLE_AIP_44", None, "true")
         _set_var(_env, "AIRFLOW_ENV", "development")
         _set_var(_env, "AIRFLOW_EXTRAS", self.airflow_extras)
+        _set_var(_env, "AIRFLOW_SKIP_CONSTRAINTS", self.airflow_skip_constraints)
         _set_var(_env, "AIRFLOW_IMAGE_KUBERNETES", self.airflow_image_kubernetes)
         _set_var(_env, "AIRFLOW_VERSION", self.airflow_version)
         _set_var(_env, "AIRFLOW__CELERY__BROKER_URL", self.airflow_celery_broker_url)
@@ -435,6 +493,7 @@ class ShellParams:
         _set_var(_env, "BREEZE", "true")
         _set_var(_env, "BREEZE_INIT_COMMAND", None, "")
         _set_var(_env, "CELERY_FLOWER", self.celery_flower)
+        _set_var(_env, "CHICKEN_EGG_PROVIDERS", self.chicken_egg_providers)
         _set_var(_env, "CI", None, "false")
         _set_var(_env, "CI_BUILD_ID", None, "0")
         _set_var(_env, "CI_EVENT_TYPE", None, "pull_request")
@@ -475,6 +534,10 @@ class ShellParams:
         _set_var(_env, "PACKAGE_FORMAT", self.package_format)
         _set_var(_env, "POSTGRES_HOST_PORT", None, POSTGRES_HOST_PORT)
         _set_var(_env, "POSTGRES_VERSION", self.postgres_version)
+        _set_var(_env, "PROVIDERS_CONSTRAINTS_LOCATION", self.providers_constraints_location)
+        _set_var(_env, "PROVIDERS_CONSTRAINTS_MODE", self.providers_constraints_mode)
+        _set_var(_env, "PROVIDERS_CONSTRAINTS_REFERENCE", self.providers_constraints_reference)
+        _set_var(_env, "PROVIDERS_SKIP_CONSTRAINTS", self.providers_skip_constraints)
         _set_var(_env, "PYTHONDONTWRITEBYTECODE", "true")
         _set_var(_env, "PYTHONWARNINGS", None, None)
         _set_var(_env, "PYTHON_MAJOR_MINOR_VERSION", self.python)
@@ -484,7 +547,6 @@ class ShellParams:
         _set_var(_env, "REMOVE_ARM_PACKAGES", self.remove_arm_packages)
         _set_var(_env, "RUN_SYSTEM_TESTS", self.run_system_tests)
         _set_var(_env, "RUN_TESTS", self.run_tests)
-        _set_var(_env, "SKIP_CONSTRAINTS", self.skip_constraints)
         _set_var(_env, "SKIP_ENVIRONMENT_INITIALIZATION", self.skip_environment_initialization)
         _set_var(_env, "SKIP_SSH_SETUP", self.skip_ssh_setup)
         _set_var(_env, "SQLITE_URL", self.sqlite_url)
@@ -503,12 +565,11 @@ class ShellParams:
         _set_var(_env, "WEBSERVER_HOST_PORT", None, WEBSERVER_HOST_PORT)
         _set_var(_env, "_AIRFLOW_RUN_DB_TESTS_ONLY", self.run_db_tests_only)
         _set_var(_env, "_AIRFLOW_SKIP_DB_TESTS", self.skip_db_tests)
-
         self._generate_env_for_docker_compose_file_if_needed(_env)
 
-        target_environment = deepcopy(os.environ)
-        target_environment.update(_env)
-        return target_environment
+        _target_env: dict[str, str] = os.environ.copy()
+        _target_env.update(_env)
+        return _target_env
 
     @staticmethod
     def _generate_env_for_docker_compose_file_if_needed(env: dict[str, str]):
@@ -578,3 +639,9 @@ class ShellParams:
             GENERATED_DOCKER_COMPOSE_ENV_FILE.write_text(
                 "\n".join([f"{k}=${{{k}}}" for k in sorted(env.keys())])
             )
+
+    def __post_init__(self):
+        if self.airflow_constraints_reference == "default":
+            self.airflow_constraints_reference = self.default_constraints_branch
+        if self.providers_constraints_reference == "default":
+            self.providers_constraints_reference = self.default_constraints_branch