+++ dev/breeze/tests/test_packages.py
@@ -104,22 +104,22 @@ def test_get_long_package_name():
 
 def test_get_provider_requirements():
     # update me when asana dependencies change
-    assert get_provider_requirements("asana") == ["apache-airflow>=2.5.0", "asana>=0.10,<4.0.0"]
+    assert get_provider_requirements("asana") == ["apache-airflow>=2.6.0", "asana>=0.10,<4.0.0"]
 
 
 def test_get_removed_providers():
     # Modify it every time we schedule provider for removal or remove it
-    assert [] == get_removed_provider_ids()
+    assert ["apache.sqoop", "plexus"] == get_removed_provider_ids()
 
 
 def test_get_suspended_provider_ids():
     # Modify it every time we suspend/resume provider
-    assert [] == get_suspended_provider_ids()
+    assert ["apache.sqoop", "plexus"] == get_suspended_provider_ids()
 
 
 def test_get_suspended_provider_folders():
     # Modify it every time we suspend/resume provider
-    assert [] == get_suspended_provider_folders()
+    assert ["apache/sqoop", "plexus"] == get_suspended_provider_folders()
 
 
 @pytest.mark.parametrize(
@@ -163,7 +163,7 @@ def test_get_install_requirements():
     assert (
         get_install_requirements("asana", "").strip()
         == """
-    "apache-airflow>=2.5.0",
+    "apache-airflow>=2.6.0",
     "asana>=0.10,<4.0.0",
 """.strip()
     )
@@ -313,7 +313,7 @@ def test_validate_provider_info_with_schema():
 @pytest.mark.parametrize(
     "provider_id, min_version",
     [
-        ("amazon", "2.5.0"),
+        ("amazon", "2.6.0"),
         ("common.io", "2.8.0"),
     ],
 )
@@ -378,7 +378,7 @@ def test_provider_jinja_context():
         "CHANGELOG_RELATIVE_PATH": "../../airflow/providers/amazon",
         "SUPPORTED_PYTHON_VERSIONS": ["3.8", "3.9", "3.10", "3.11"],
         "PLUGINS": [],
-        "MIN_AIRFLOW_VERSION": "2.5.0",
+        "MIN_AIRFLOW_VERSION": "2.6.0",
         "PROVIDER_REMOVED": False,
         "PROVIDER_INFO": provider_info,
     }