+++ airflow/providers/docker/CHANGELOG.rst
@@ -27,6 +27,30 @@
 Changelog
 ---------
 
+3.9.0
+.....
+
+.. note::
+  This release of provider is only available for Airflow 2.6+ as explained in the
+  `Apache Airflow providers support policy <https://github.com/apache/airflow/blob/main/PROVIDERS.rst#minimum-supported-version-of-airflow-for-community-managed-providers>`_.
+
+Bug Fixes
+~~~~~~~~~
+
+* ``Fix 'enable_logging=True' not working in 'DockerSwarmOperator' (#35677)``
+* ``Fix broken log streaming from #35677 (#36127)``
+
+Misc
+~~~~
+
+* ``Bump minimum Airflow version in providers to Airflow 2.6.0 (#36017)``
+* ``Follow BaseHook connection fields method signature in child classes (#36086)``
+
+.. Below changes are excluded from the changelog. Move them to
+   appropriate section above if needed. Do not delete the lines(!):
+   * ``Update information about links into the provider.yaml files (#35837)``
+   * ``Prepare docs 1st wave of Providers December 2023 (#36112)``
+
 3.8.2
 .....